Add in EVERYTHING
This should work at least for CCS.  Project files are in place for all the other
builds, and the makefile's in there for GCC, but the library paths aren't
verified for anything other than CCS at the moment.  I'm guessing if you tried
to build it in another environment, it would bork, but would be easy to fix by
manually adding a library path to the cmsis.lib for that compiler.